feat: add NewTransactionModal component and styles
This commit introduces the NewTransactionModal component with relevant styles for creating new transactions. In addition, it includes updates on the package lock file to add new dependencies required. '@phosphor-icons/react', '@radix-ui/react-dialog' and custom styles are used within this component to create a form for transaction attributes such as description, price and category.